agl-container-demo: enable LIN, Radio and Audio
For agl ivi features, the kernel modules of LIN and Audio and
firmware of radio should be installed into host kernel and rootfs.

And to complete security setting of pipewire, it should be
installed so far (it might be caused by sharing SMACK between
host and guest to simplify demo integration.
To workaround this, pipewire is installed into host rootfs.

Change-Id: I0bd4af9c3c9eb73e4f18ae592adaf6dca0b65930
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>