agl-container-demo: Remove title and border
When using wayland-backend, due to the addition of borders and title,
the screen of nested compositor shifts down to the lower right than
exptected and is also out of sight.

Originally, --fullscreen option is supported to serve this purpose,
but this causes crush (coredump).

As workaround, this patch fixes libweston directly.

Original commit in meta-agl:sandbox/harunobu.kurokawa/ces_2020_debug1231
    commit dc05cade60cebae8d7d4401e2e7b78c81b10ba96
    Author: Harunobu Kurokawa <harunobu.kurokawa.dn@renesas.com>
    Date:   Mon Nov 25 18:30:07 2019 +0900

    weston: Disable wayland backend frame

    Change-Id: I94204c204a774d66e087e636c1bd10e27b643d9f
    Signed-off-by: Harunobu Kurokawa <harunobu.kurokawa.dn@renesas.com>

Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>
Change-Id: I6bafc855c23506c506f9299a0372feff55681f47