meta-agl-basesystem: add libxml-xpath-perl
Add libxml-xpath-perl recipe and recipes for its required dependencies
libscalar-list-utils-perl and libpath-tiny-perl.  This should enable
using the xpath utility in the basesystem build by DEPENDing on
libxml-xpath-perl-native.

Note that the libscalar-list-utils-perl recipe has been adapted from
the one present in meta-cpan (https://github.com/meta-cpan/meta-cpan)
as of commit 2f6ecfa.  This was viewed as preferable to pulling in
meta-cpan as an AGL dependency.  Additionally, the recipe has been
updated to enable ptest support, and all tests have been confirmed to
pass.  The same is true of the libxml-xpath-perl and libpath-tiny-perl
ptests.

Bug-AGL: SPEC-3670

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Icd25cbb4498636c5a6a384818535e3c3372b5b29