drm-lease: Update drm-lease-manager
Adds option to keep a lease alive if its client exits unexpededly.
This will stop the screen from blanking if the client crashes or
otherwise shuts down abnormally while holding a lease.  The
screen will blank as normal if a client explicitly releases the lease.

This option is enabled or disabled based on the command line arguments
given to the DRM lease manager at runtime.

Change-Id: I2e80c1cb1b31be756c8ba8e4e9fba1adfcc75447