drm-lease: RPi4: Disable DRM_AUTH checks on wayland_clients
DRM authentication does not work via a DRM Leases because the client
can only authenticate against a lease owner (primary node), since
the lease does not have a corresponding device file in the filesystem
for clients to open.

Bypassing DRM authentication by using render nodes is not yet fully
supported in mesa, so for now, just try to authenticate and
continue even if it fails.

Bug-AGL: SPEC-4317

Change-Id: Ie6c82162037964da697d3ece80a8c463980d2658
Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>