lxc-config-ivi-demo-platform: AGL compositor basic image
This patch adds will create a new LXC container image with the multi
configuration environment available. The image contains the AGL
compositor and with it, the necessary bits to have user session working.

Of particular importance is the fact that for user sessions to work we
require having tty0 present in the container image and permission to
change from text mode to graphics mode using systemd-logind inside the
container image.

This patch contains the LXC configuration to handle that.

Bug-AGL: SPEC-4606
Change-Id: I2d7f3338418ff05cab6c7b5b91555622f0f52ab8
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>